+++ app/_components/Auth/Navigate.tsx
@@ -1,0 +1,15 @@
+import { useRouter, useSearchParams } from 'next/navigation';
+import { useEffect } from 'react';
+
+export default function Navigate() {
+  const router = useRouter();
+  const goto = useSearchParams().get('goto') || 'mydashboard';
+
+  useEffect(() => {
+    if (typeof window !== undefined && goto && router) {
+      router.push(goto);
+    }
+  }, [goto, router]);
+
+  return null;
+}
